pop up changed as per requirements
